requirements.txt
customtkinter~=5.2.2
pillow~=10.2.0
pyperclip~=1.8.2